Fix error on prop key and refactor component
